#2449 Fix Intel build.
git-svn-id: https://chaste.cs.ox.ac.uk/svn/chaste/trunk@20521 b5f69653-2bf5-0310-8201-b0dafb8e1405

+++ io/src/writer/Hdf5DataWriter.cpp
@@ -1192,7 +1192,7 @@ void Hdf5DataWriter::SetChunkSize()
         // 128 M by default which should be plenty.
         unsigned target_size_bytes = 1024*1024/8;
         unsigned target_size = 0;
-        unsigned divisors[DATASET_DIMS];
+        long unsigned divisors[DATASET_DIMS];
         unsigned chunk_size_bytes = 0;
 
         while ( chunk_size_bytes < target_size_bytes )
